+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Blueprints for armor, elytra, and shields now display their armor durability modifier [#521]
+
 ## [2.10.7] - 2022-05-30
 ### Fixed
 - Imperial and Gold Digger traits not working [#509]